feature(mapper): Add logic for instanced areas (#46)
* Adds logic for instanced areas

Calls to deleteArea when switching to a new zone if that zone is "instanced".  There may be instances where this is incorrect, but I haven't run into any since adding the logic for ignoring "Battlefield".

* Handle case where map.prev_info.area is not yet set

This likely would have led to the deletion of the area a user logged into.